Update notifier_bot.py
changed Tracking Sent Posts:

A file (sent_posts.txt) is used to store the URLs of job posts that have already been sent. This prevents the same post from being sent multiple times.
Loading and Saving Sent Posts:

load_sent_posts() function loads the URLs of the posts that have already been sent.
save_sent_posts() function updates the file with new posts that have been sent.
Avoiding Duplicate Emails:

Before sending an email, the script checks if a job post's URL is already in the sent_posts.txt file. If it is, that post will not be included in the email.
Storing Sent URLs:

After sending the email, the URLs of the newly sent posts are added to the sent_posts.txt file to keep track of the posts that have been sent.